Change Locations and Zones columns in DB

+++ src/esMWS.Domain/Entities/WarehouseEnviroment/Zone.cs
@@ -6,7 +6,7 @@
         public string? ZoneName { get; set; }
         public char ZoneAlias { get; set; }
         public string WarehouseId { get; set; } = null!;
-        public int? AvgTemperature { get; set; }
+        public decimal? AvgTemperature { get; set; }
         public DateTime CreatedAt { get; set; }
         public string? CreatedBy { get; set; }
         public DateTime? ModifiedAt { get; set; }
